requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0
Werkzeug>=2.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
pytesseract>=0.3.10

